Enable Sentry frontend FullStory integration
If FullStory is enabled for the current session, this integration will
associate the FullStory session URL with any errors sent to Sentry.
Additionally, it will add an event to the FullStory session indicating
there was a Sentry error, and providing a URL for easy access. If
FullStory is not enabled for the current session, this integration is not
enabled.

closes DE-996
flag=none

test plan:
- ensure that the appropriate values are provided in config/sentry.yml
- enable FullStory for a session by overriding methods in FullStoryHelper
- cause an unhandled exception to occur on the frontend
- verify that the event in Sentry contains FullStory context with a link
  to the session URL
- verify that the FullStory session contains the Sentry Error event, with
  a link to the Sentry event
- disable FullStory, and verify that frontend exception monitoring with
  Sentry works as normal

Change-Id: I782930ecbea6f3e4afb23045d4087fe4c252875d
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/283151
QA-Review: Isaac Moore <isaac.moore@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>