babel independence(3): jest
refs FOO-2697
flag = none

on "babel-plugin-transform-imports": the behavior of this plugin was
surprising to me in that it not only rewrote the specifiers to point to
the CommonJS instead of the ESM variants, but it actually changed the
specifiers to point to the exact file that exports the symbol through a
mixture of fs globbing and guesses.

Consider the following import:

    import { DateTime } from '@instructure/ui-i18n'

Normally, you'd think we'll be loading the {main,module,browser} file of
the @instructure/ui-i18n package and refer to the `DateTime` symbol
exported by that file. With the plugin activated, however, the import
*itself* changes to:

    import DateTime from '@instructure/ui-i18n/lib/DateTime.js'

This is no longer merely a variant rewrite in any way as we're no longer
loading the same file. This caused new problems to surface that up to
this point were not encountered, as now we're actually loading the
entrypoints as specified.

An example of this is in that specific package where it (its entrypoint)
loads a package `decimal.js` whose ESM variant doesn't work as-is but
its CJS variant does -- something neither we nor the InstUI team were
privy to as the plugin was simply clobbering this piece of information.

~ test plan ~

- jest suite runs and passes

Change-Id: I55d7db1ede742e0f588a84d54a761b7aba03729f
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/283543
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Aaron Shafovaloff <ashafovaloff@instructure.com>
QA-Review: Aaron Shafovaloff <ashafovaloff@instructure.com>
Product-Review: Ahmad Amireh <ahmad@instructure.com>