Restrict edit option when course is concluded
fixes MAT-510
flag=none

Test Plan:
- Create a course and add a new
teacher in it. Teacher must
have accepted the invitation for
the course
- Once listed, create 3 pages in
the course with the following
editing options:

* page 1 'Only teachers'
* page 2 'Teachers and students'
* page 3 'Anyone'

- Save and publish each new page
- When clicking on each page, the 'Edit'
option is there. All ok.
- Go to 'Settings' and click on
'Conclude the course'. Conclude it.
Just in case, keep in mind the ID
of the concluded course to check
the pages
- When you locate the pages URL,
masquerade as a teacher and check
on each page. The 'Edit' option is
no longer shown.

Change-Id: I05be72facf0155922a52891dc34aa7bf6e560e7a
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/284681
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Weston Dransfield <wdransfield@instructure.com>
QA-Review: Weston Dransfield <wdransfield@instructure.com>
Product-Review: David Lyons <lyons@instructure.com>