B&I: Make button name textbox reflect the actual file name
We're getting rid of the concept of "display name" in a B&I; instead the
"Name" textbox will just show the name of the file. For new buttons it
will be used to determine the filename, but we'll still use Files' logic
for appending suffixes (-1, -2, etc) when there are name collisions, and
the textbox will reflect that. We'll disable the "Name" textbox when we
are doing a "Replace all" since we don't want to change the filename in
that operation.

refs MAT-712
flag=buttons_and_icons_root_account

test plan

Name text box should be disabled when replacing all:
- Insert a button
  Take note of the filename
- Edit the button
- Change the button's name
- Make some other changes
- Scroll down and tick the "apply to all" checkbox
  CONFIRM that the button's name has reverted back,
  and the name text box is disabled
- Save
- Go to Files
- Find the button you were editing
  CONFIRM that it was updated with the changes you made above
  CONFIRM that no new buttons were created when you Saved

Name text box should reflect the actual file name of the button:
- Create a button, leaving the name blank
- Apply
- Edit your button
- Observe that the name now is "untitled"
  (or maybe untitled-1, untitled-2, etc depending on how many
  untitled B&I you lready had)

Name text box should reflect the file name even if it has changed:
- Create a button and set a name
- Apply
- Edit your button
- Observe that the name is still the same
- Click save with the "apply all" checkbox UNCHECKED
- Edit your button again
- Observe that the name now has a `-1` at the end

Name is not embedded in SVG:
- Download one of the buttons you just created and open it in a text
  editor
- Verify that the SVG metadata section doesn't contain the button's name

Change-Id: Ice7af262d1eeaab1612d76a11c5490ddf6bfcd9f
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/286201
Reviewed-by: Juan Chavez <juan.chavez@instructure.com>
Reviewed-by: Gonzalo Penaranda <gonzalo.penaranda@instructure.com>
QA-Review: Gonzalo Penaranda <gonzalo.penaranda@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: David Lyons <lyons@instructure.com>