Revert unnecessary card image resize from LS-1122
For LS-1122 I down-sampled the image selected for the course
card before uploading. Because this image is also used as the
default banner image for k5 courses. This had the side effect
of making that image blurry. It turns out this resize was
unnecessary since DashboardCard adds a geometry parameter to
the image URL which has inst-fs do the resizing for us.

In retrospect, I don't think the issue described in LS-1122
was true cause-and-effect and have asked the reporter to
investigate further.

This PS reverts the changes from LS-1122 and g/287583.

closes LS-3166
flag=none

test plan:
  To _really_ test you need your canvas using inst-fs
  but you could:
  - add a card image to your course
  - visit the dashboard
  - inspect the course card's CSS
  - download the image from the URL specified in the
    card's background-image
  > expect to see that it's the same size as the
    one you uploaded
  - repeat for a k5 course
  - visit the course page
  > expect the banner image to look good (not a
    very objective test, but what are you gonna do?)

  You can also
  - repeat in beta or prod (sadly, cd isn't using inst-fs)
  - inspect the card's background-image CSS
  - see that it has geometry=widthxheight in the url

  If you _do_ have inst-fs
  - upload a big card image
  - inspect the course card's CSS
  - hover over the background-image url
  > expect the file size to be smaller than the image
    you uploaded
  - or click on the URL to download the rendered image and
  > expect the image and file size to be smaller than the image
    you uploaded

Change-Id: I9d19acb062f49ee82e351ac29dd5359cea3c0b25
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/291658
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jackson Howe <jackson.howe@instructure.com>
QA-Review: Jackson Howe <jackson.howe@instructure.com>
Product-Review: Allison Howell <allison.howell@instructure.com>