Put correct assignment icons into the course syllabus
The calendar_events api returned type="assignment" for anything
that's an assignment, including graded discussions and quizzes.
This PS adds "submission_types" to the returned json so the syllabus
can distinguish between the sub-types.
FYI, this is not an issue in the calendar because there the
calendar_events api is told to return the assignment, so it has
enough info to do the job.  (It's possible that returning the
submission_types would be sufficient there too)

closes LS-591
flag=none

test plan:
  - in a course create
    - a discussion and check "add to student to do" and give it a date
    - a graded discussion
    - a page and make it a to do also
    - a quiz
    - an assignment
    - a calendar event
  - go to the syllabus page
  > expect the 6 items to be there, each with the correct icon
  > expect the screenreader to correctly say what each of them are

Change-Id: I35eb312b74d3cff3b308789390e0175baf5439ed
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/291686
Reviewed-by: Robin Kuss <rkuss@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Allison Howell <allison.howell@instructure.com>