Fix: different AG conversations have the same thread
The context code wasn't being included when sending messages which
caused all the messages to be associated to the conversation with the
context_id nill, this PS makes the context_code be included in the
send message payload, so the conversation can be initialized properly

closes: LS-3144
flag= none

Test plan:

Prerequisite: Two published courses with the same student and teacher
in both

- As the teacher, create an appointment group in each course.
- In one of the appointment groups, use the Message Students option
and send a message to the student.
- Send a message using the same method in the appointment on the
other course.
- Navigate to Inbox, filter Sent messages, and expect to see 2
different conversations, one for each course

Change-Id: I80f59f5fdb00efa9588602a0b1d1e4bf996ca84c
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/291853
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Jonathan Guardado <jonathan.guardado@instructure.com>
Reviewed-by: Robin Kuss <rkuss@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>