Wire up the new InstUI Create New Groupset modal
Fixes FOO-906
flag=none

[ignore-stage-results=Flakey Spec Catcher]

Replaces the old Coffeescript and Handlebars code for creating a
new group set with a set of React components that use InstUI for
a more accessible and responsive experience.

This commit skips the flaky spec catcher because of a single spec
taking too long to run:
	(spec/selenium/groups/manage_new_groups_spec.rb:32)
I've run this spec repeatedly by hand and while it does take a
while to run, it's also doing a lot (has to fire off a job and
wait for the result, etc). And it passes for me 100% of the time,
so rather than skip the spec itself (it seems a useful one), it's
probably best just to skip the FSC this time and move on. FSC did
not have any issues about any other specs.

Test plan:
* Try to create a new group set in an account (/accounts/:id/groups)
* You should get a nice-looking InstUI modal dialog which has all
  the functionality of the old HTML one. This is a much simpler version
  of the modal that only lets you choose the name of a new group set.
* Now try to create a new group set in a course (/courses/:id/groups)
* You should get a new InstUI model which has all the functionality of
  the old one. This is the more complex version of the modal with lots
  of options and checkboxes and fields to fill out.
* When the form is properly filled out, clicking "Save" should actually
  create the group set and it should show up in the tabs.
* Try filling the form out incorrectly (no name, negative numbers in the
  count fields, etc.
* With incorrect filling-out, clicking "Save" should disable the Save
  button and point out the error with red highlights and error text.

Change-Id: I768ff665bf98d2661e06f46480cc762a73b613d6
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/293258
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: August Thornton <august@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>